Create build-entrypoint.sh to handle first stage of container based-build
This file will be the first point of entry in the docker container as an MKP is built. It copies the source from the mapped folder and triggers any updates needed. It also triggers the build-modify-extension.py which will add appropriate data to the MKP manifest and any other files in the mkp before it is built  (eg. Add CMK_VERSION= to the bash agent script